feat(installation): enhance installer with routing and debugbar integration
- Added routing for installation steps in `bootstrap/app.php`.
- Removed `RouteServiceProvider` from `providers.php`.
- Integrated `laravel-debugbar` for improved debugging in development.
- Created `debugbar.php` configuration file.
- Updated migration for redirects to use `unsignedSmallInteger` for status codes.
- Modified installation views to use asset helper for CSS.
- Enhanced database configuration view with AJAX connection testing.
- Added CSRF token meta tag for AJAX requests in installation views.
- Improved form structure and validation feedback in installation views.
- Updated routes to streamline installation process.
- Added `.gitignore` for debugbar storage directory.

+++ database/migrations/2026_08_11_170001_create_redirects_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->string('from_url')->unique();
             $table->string('to_url');
-            $table->unsignedTinyInteger('status_code')->default(301);
+            $table->unsignedSmallInteger('status_code')->default(301);
             $table->timestamps();
         });
     }
